requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
requests>=2.26.0
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.0
pyarrow>=7.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.10.0

# Monitoring
prometheus-client>=0.14.0
grafana-api>=1.0.3

# Workflow
apache-airflow>=2.5.0
apache-airflow-providers-postgres>=5.0.0

# Development
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
mypy>=0.910

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# New dependencies
pyspark==3.4.1
fastparquet==2023.10.1
beautifulsoup4==4.12.2
selenium==4.15.2

# Database connectors
pymongo==4.5.0
mysql-connector-python==8.1.0

# Development tools
black==23.10.1
flake8==6.1.0
isort==5.12.0
mypy==1.6.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# New dependencies
dask==2023.12.1
sqlalchemy==2.0.23
docker==7.0.0
